add mesh

+++ src/cs4620/splines/CatmullRom.java
@@ -23,8 +23,8 @@ public class CatmullRom extends SplineCurve {
 		//SOLUTION
 		Matrix4 m = new Matrix4(
 				0, 1, 0, 0,
-				- 1 / 6, 1, 1 / 6, 0,
-				0, 1 / 6, 1, - 1 / 6,
+				(float)- 1 / 6, 1, (float)1 / 6, 0,
+				0, (float)1 / 6, 1, (float)- 1 / 6,
 				0, 0, 1, 0
 				);
 		Vector4 x = m.mul(new Vector4 (p0.x, p1.x, p2.x, p3.x));
@@ -32,4 +32,15 @@ public class CatmullRom extends SplineCurve {
 		return new CubicBezier(new Vector2(x.x, y.x), new Vector2(x.y, y.y), new Vector2(x.z, y.z), new Vector2(x.w, y.w), eps);
 		//END SOLUTION
 	}
+//	 public static void main(String[] args) {
+//	    	Vector2 p0 = new Vector2(0,0);
+//	    	Vector2 p1 = new Vector2(1,2);
+//	    	Vector2 p2 = new Vector2(3,2);
+//	    	Vector2 p3 = new Vector2(4,0);
+//	    	ArrayList<Vector2> controlPoints = new ArrayList<Vector2>();
+//	    	controlPoints.add(p0);controlPoints.add(p1);controlPoints.add(p2);controlPoints.add(p3);
+//	    	CubicBezier cb = new CatmullRom(controlPoints,false,(float)0.1).toBezier(p0, p1, p2, p3,(float)0.1);
+//	    	System.out.println();
+//	    	
+//		}
 }
